njit temperature cost functional

+++ conmech/vertex_utils.py
@@ -7,4 +7,4 @@ import numpy as np
 
 @numba.njit(inline='always')
 def length(p1, p2):
-    return float(np.sqrt((p1[0] - p2[0]) ** 2 + (p1[1] - p2[1]) ** 2))
+    return np.sqrt((p1[0] - p2[0]) ** 2 + (p1[1] - p2[1]) ** 2)
